Add support for file buffers (#50)
<!-- Provide a general summary of your changes in the Title above -->

## Description
Previously, `parsnip` expected a path to a file (provided as either a
str or a pathlib.Path), which was opened into a file buffer in the
__init__ method. This is not optional, and a direct file buffer can now
be provided. We also support raw text blocks (str containing newlines),
which are preprocessed appropriately.
**We make the assumption that all strings that do not point to a file,
do not contain a portion of a valid path, and do not contain the ".cif"
suffix should be treated as CIF data blocks.**

## Motivation and Context

Resolves #49
Blocked by #51 

## Types of Changes
<!-- Please select all items that apply, either now or after creating
the pull request: -->
- [ ] Documentation update
- [ ] Bug fix
- [x] New feature
- [ ] Breaking change<sup>1</sup>

<sup>1</sup>The change breaks (or has the potential to break) existing
functionality and should be merged into the `breaking` branch

## Checklist:
<!-- Please select all items that apply either now or after creating the
pull request. -->
- [x] I am familiar with the [Development
Guidelines](https://github.com/glotzerlab/parsnip/blob/main/doc/source/development.rst)
- [x] The changes introduced by this pull request are covered by
existing or newly introduced tests.
- [x] I have updated the
[changelog](https://github.com/glotzerlab/parsnip/blob/main/ChangeLog.rst)
and added my name to the
[credits](https://github.com/glotzerlab/parsnip/blob/main/doc/source/credits.rst).